Add a note to an existing RData file

Use this to add a note to a saved R data file that was created with digitizeRadii. Can be used to flag the reading for some reason.

addNote(nms, note)

Arguments

nms

A string that indicates the R data file created with digitizeRadii (in the current directory) to which the note should be added.

note

The note (as a string) to be added to the R data file. If missing the user will be provided a dialog box in which to enter the note.

Value

None, but the note object in the R Data file(s) given in nms will be modified.

Details

A detailed description of its use is in the "Other Features" vignette on the RFishBC website.

Author

Derek H. Ogle, derek@derekogle.com

Examples

## See the link to the extensive documentation in the Details.