Collect radial measurements from a calcified structure by interactively selecting annuli.

Computes radial measurements from selected points and writes all to an R data object

The user can interactively select points on a plot by clicking with the first mouse button on the image. When finished, the x- and y-coordinates for each selected point will be returned.

digitizeRadii(img, id, reading, suffix, description, edgeIsAnnulus, popID,
  sepWindow, windowSize, scaleBar, scaleBarLength, col.scaleBar, lwd.scaleBar,
  scalingFactor, addTransect, snap2Transect, col.transect, lwd.transect,
  pch.sel, col.sel, cex.sel, showInfo, pos.info, cex.info, col.info)

Arguments

img
A string that indicates the image (must be PNG, JPG, or BMP) to be loaded and plotted. By default the user will be provided a dialog box from which to choose the file. Alternatively the user can supply the name of the file (will look for this file in the current working directory unless a fully pathed name is given).

id
A unique identifier for the fish or structure being examined. Will be coerced to a character.

reading
See details in RFBCoptions.

suffix
See details in RFBCoptions.

description
See details in RFBCoptions.

edgeIsAnnulus
See details in RFBCoptions.

popID
See details in RFBCoptions.

sepWindow
See details in RFBCoptions.

windowSize
See details in RFBCoptions.

scaleBar
See details in RFBCoptions.

scaleBarLength
See details in RFBCoptions.

col.scaleBar
See details in RFBCoptions.

lwd.scaleBar
See details in RFBCoptions.

scalingFactor
See details in RFBCoptions.

addTransect
See details in RFBCoptions.

snap2Transect
See details in RFBCoptions.

col.transect
See details in RFBCoptions.

lwd.transect
See details in RFBCoptions.

pch.sel
See details in RFBCoptions.

col.sel
See details in RFBCoptions.

cex.sel
See details in RFBCoptions.

showInfo
See details in RFBCoptions.

pos.info
See details in RFBCoptions.

cex.info
See details in RFBCoptions.

col.info
See details in RFBCoptions.

Value

A string that contains the filename (with path) of an R data object (i.e., .RData file) written to the same directory/folder as the file in fname that contains the following items in a list:

description: The description given in description.

id: The unique identifier given in id.

image: The filename for the structure image given in fname.

pts: The data.frame of x- and y-coordinates for the points (as returned directly from selecting points on the structure image).

radii: A data.frame that contains the radial measurements (i.e., from the focus to the selected point) for the structure. Specifically, it contains the following items:

XXX

.

Details

This uses locator which will suspend activity in the console until the user indicates that they are done selecting points. The user indicates that they are done selecting points by right-clicking on the image and selecting Stop or selecting Stop in the window's menu.

Examples

## None yet