Returns the notes from the chosen R Data files

Returns the notes from the chosen R data files created with digitizeRadii.

findNotes(nms)

Arguments

nms

A string (or vector of strings) that indicates the R data file(s) created with digitizeRadii. If missing the user will be provided a dialog box from which to choose the file(s). The file(s) must be in the current working directory (see getwd result). May also be a single RFishBC object created with digitizeRadii.

Value

A data.frame that contains the id, reading, and notes from the chosen R Data files.

Details

A detailed description of its use is in the "Other Features" vignette on the RFishBC website.

Author

Derek H. Ogle, derek@derekogle.com

Examples

## See the link to the extensive documentation in the Details.