Find scaling factor from object of known length

This computes a scaling factor (i.e., a multiplier that is used to convert image lengths to actual lengths) from user selected endpoints of an object of known length on the image given in img and the actual known length given in knownLength. See Details.

findScalingFactor(
  img,
  knownLength,
  windowSize,
  deviceType,
  closeWindow,
  col.scaleBar,
  lwd.scaleBar,
  pch.sel,
  col.sel,
  cex.sel,
  pch.del,
  col.del
)

Arguments

img

A string that indicates the image (must be PNG, JPG, BMP, or TIFF) to be loaded and plotted. By default the user will be provided a dialog box from which to choose the file. Alternatively the user can supply the name of the file (will look for this file in the current working directory).

knownLength

See details in RFBCoptions.

windowSize

See details in RFBCoptions.

deviceType

See details in RFBCoptions.

closeWindow

See details in RFBCoptions.

col.scaleBar

See details in RFBCoptions.

lwd.scaleBar

See details in RFBCoptions.

pch.sel

See details in RFBCoptions.

col.sel

See details in RFBCoptions.

cex.sel

See details in RFBCoptions.

pch.del

See details in RFBCoptions.

col.del

See details in RFBCoptions.

Value

A single numeric that is the scaling factor (a multiplier that is used to convert image lengths to actual lengths).

Details

To apply the scaling factor determined with this function to images opened in digitizeRadii is is important that the images were created with the EXACT same magnification, are saved with the EXACT same dimensions (and aspect ratio), and the EXACT same value for windowSize= is used.

Author

Derek H. Ogle, derek@derekogle.com

Examples

## None yet