Radial measurements for for West Bearskin Lake Smallmouth Bass.

Radial measurement data for West Bearskin Lake Smallmouth Bass. Data are in “wide” or one-fish-per-line (with radial measurements in separate variables) format.

SMBassWB2

Format

A data frame with 181 rows and 13 variables:

id

An identification number unique to each fish.

reading

A reading specific identifier (all are “DHO”).

agecap

Age-at-capture.

radcap

Total structure radius at capture (magnified mm).

rad1

Structure radius to annulus 1.

rad2

Structure radius to annulus 2.

rad3

Structure radius to annulus 3.

rad4

Structure radius to annulus 4.

rad5

Structure radius to annulus 5.

rad6

Structure radius to annulus 6.

rad7

Structure radius to annulus 7.

rad8

Structure radius to annulus 8.

rad9

Structure radius to annulus 9.

Source

Originally from SMBassWB in FSA. Rearranged to demonstrate functions in this package.

See also