Creates a function for a specific model.

Creates a function for a specific model based on definitions in Vigliola and Meekan (2009).

bcFuns(BCM)

Arguments

BCM
A single numeric between 1 and 22 or a string that indicates which model to use (based on numbers and names in Vigliola and Meekan (2009)).

Value

A function that can be used to predict length at previous age (Li) given length-at-capture (Lcap), hard-part radius-at-age i (Ri), and hard-part radius-at-capture (Rcap). In addition, some functions/models may require the previous age (agei) and the age-at-capture (agec), certain parameters related to the biological intercept (R0p & L0p), or certain parameters estimated from various regression models (a,b,c,A,B,C). See source for more information.

Details

The following models, based on definitions with abbreviations and model numbers from Vigliola and Meekan (2009), are supported.

Abbreviation	Number	Model

DALE	1	Dahl-Lea

FRALE	2	Fraser-Lee

BI, LBI	3	(Linear) Biological Intercept

BPH, LBPH	4	(Linear) Body Proportional Hypothesis

TVG	5	Time-Varying Growth

SPH, LSPH	6	(Linear) Scale Proportional Hypothesis

AE, AESPH	7	(Age Effect) Scale Proportional Hypothesis

AEBPH	8	(Age Effect) Body Proportional Hypothesis

MONA	9	Monastyrsky

MONA-BPH	10	Monastyrsky Body Proportional Hypothesis

MONA-SPH	11	Monastyrsky Scale Proportional Hypothesis

WAKU	12	Watanabe and Kuroki

FRY	13	Fry

MF, ABI	14	Modified Fry, Allometric Biological Intercept

FRY-BPH, ABPH	15	Fry, Allometric Body Proportional Hypothesis

FRY-SPH, ASPH	16	Fry, Allometric Scale Proportional Hypothesis

QBPH	17	Quadratic Body Proportional Hypothesis

QSPH	18	Quadratic Scale Proportional Hypothesis

PBPH	19	Polynomial Body Proportional Hypothesis

PSPH	20	Polynomial Scale Proportional Hypothesis

EBPH	21	Exponential Body Proportional Hypothesis

ESPH	22	Exponential Scale Proportional Hypothesis

Abbreviation	Number

IFAR Supplement

http://derekogle.com/IFAR/supplements/backcalculation/

References

Vigliola, L. and M.G. Meekan. 2009. The back-calculation of fish growth from otoliths. pp. 174-211. in B.S. Green et al. (editors). Tropical Fish Otoliths: Information for Assessment, Management and Ecology. Review: Methods and Technologies in Fish Biology and Fisheries 11. Springer. [Was (is?) available from https://www.researchgate.net/publication/226394736_The_Back-Calculation_of_Fish_Growth_From_Otoliths.]

Examples

## Simple Examples
( bcm1 <- bcFuns(1) )

#> function(Lcap,Ri,Rcap,...,verbose=TRUE) {
#>         if (verbose) DONE("Using the Dahl-Lea model.")
#>         (Ri/Rcap)*Lcap
#>       }
#> <environment: 0x598c6b0>

bcm1(20,10,40)

#> ✔ Using the Dahl-Lea model.

#> [1] 5

## Example with dummy length-at-cap, radii-at-cap, and radii-at-age
lencap <- c(100,100,100,150,150)
radcap <- c(20,20,20,30,30)
rad    <- c( 5,10,15,15,25)
bcm1(lencap,rad,radcap)

#> ✔ Using the Dahl-Lea model.

#> [1]  25  50  75  75 125

( bcm2 <- bcFuns("FRALE") )

#> function(Lcap,Ri,Rcap,a,...,verbose=TRUE) {
#>         if (verbose) DONE("Using the Fraser-Lee model with a=",a,".")
#>         a+(Lcap-a)*(Ri/Rcap)
#>       }
#> <environment: 0x6234100>

bcm2(lencap,rad,radcap,2)  # demonstrated with a=2

#> ✔ Using the Fraser-Lee model with a=2.

#> [1]  26.5000  51.0000  75.5000  76.0000 125.3333