Show points selected on a structure image and saved in an R data file

Show points selected on a structure image to represent annuli that were saved to an R data file using digitizeRadii. This allows the user to reexamine the selected points or overlay selected points from multiple readings of the structure.

showDigitizedImage(
  nms,
  deviceType,
  pch.show,
  col.show,
  cex.show,
  connect,
  col.connect,
  lwd.connect,
  col.scaleBar,
  lwd.scaleBar,
  showScaleBarLength,
  cex.scaleBar,
  showAnnuliLabels,
  annuliLabels,
  col.ann,
  cex.ann,
  offset.ann
)

Arguments

nms

A string (or vector of strings) that indicates the R data file(s) created with digitizeRadii. If missing the user will be provided a dialog box from which to choose the file(s). The file(s) must be in the current working directory (see getwd result). May also be a single RFishBC object created with digitizeRadii.

deviceType

See details in RFBCoptions.

pch.show

See details in RFBCoptions.

col.show

See details in RFBCoptions.

cex.show

See details in RFBCoptions.

connect

See details in RFBCoptions.

col.connect

See details in RFBCoptions.

lwd.connect

See details in RFBCoptions.

col.scaleBar

See details in RFBCoptions.

lwd.scaleBar

See details in RFBCoptions.

showScaleBarLength

See details in RFBCoptions.

cex.scaleBar

See details in RFBCoptions.

showAnnuliLabels

See details in RFBCoptions.

annuliLabels

See details in RFBCoptions.

col.ann

See details in RFBCoptions.

cex.ann

See details in RFBCoptions.

offset.ann

See details in RFBCoptions.

Value

A list that contains the size of the window in windowSize and the pixel width to height ratio in pixW2H. In addition, an image is plotted with, at least, the selected points.

Details

This function requires interaction from the user. A detailed description of its use is in this vignette on the RFishBC website.

See also

Author

Derek H. Ogle, derek@derekogle.com

Examples

## None because this requires interaction from the user.
## See the link to the extensive documentation in the Details.